Refactor predicates to their own folder in Person model
