+++ src/test/java/staffconnect/model/person/PersonHasFacultyPredicateTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
+import staffconnect.model.person.predicates.PersonHasFacultyPredicate;
 import staffconnect.testutil.PersonBuilder;
 
 public class PersonHasFacultyPredicateTest {
