+++ src/test/java/staffconnect/model/person/PersonHasModulePredicateTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
+import staffconnect.model.person.predicates.PersonHasModulePredicate;
 import staffconnect.testutil.PersonBuilder;
 
 public class PersonHasModulePredicateTest {
